requirements.txt
opencv-python~=4.5.5.62
numpy~=1.22.1
matplotlib
Django~=4.0.2
Pillow~=9.0.1